requirements.txt
# Core Dependencies
requests>=2.31.0
pyyaml>=6.0.1
pandas>=2.0.0
numpy>=1.24.0

# ML Pipeline
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# FastAPI Backend (HIPAA-Compliant)
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Security & Encryption
cryptography>=41.0.0

# Ollama Integration
httpx>=0.25.0

# Data Validation
pydantic>=2.5.0

# Audit Logging
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
